requirements.txt
flask==3.0.2
flask-socketio==5.3.6
flask-cors==4.0.0
python-dotenv==1.0.1
openai==1.12.0
langchain==0.1.9
langchain-openai==0.0.7
transformers==4.37.2
opencv-python==4.9.0.80
PyAudio==0.2.14
SpeechRecognition==3.10.1
eventlet==0.35.1
